+++ problems/euler_001.hs
